requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
idna==2.10
pytelegrambotapi==3.7.4
python-dotenv==0.15.0
requests==2.25.0
soupsieve==2.0.1; python_version >= "3.0"
urllib3==1.26.2
wikipedia==1.4.0
